POST fixes v1.1

+++ src/components/Nfcs/WriteTag/DeleteTagInDb.jsx
@@ -4,7 +4,9 @@ export default async function deleteTagInDb(id){
     var checkBox = document.getElementById("batchCheck");
     if ("NDEFReader" in window) {
         if (checkBox.checked == false){
-            await axios.delete('https://node-nfc-db.onrender.com/api/nfcs/' + id);
+            if(id != ""){
+                await axios.delete('https://node-nfc-db.onrender.com/api/nfcs/' + id);
+            }
         }
     }
 }
